Move symbols.ts to Review dir
